scripts/intelligence/requirements.txt
# Python dependencies for Oeil de Lynx intelligence gathering

# Supabase client
supabase==2.10.0

# Web scraping
duckduckgo-search==6.3.5
youtube-search-python==1.6.6
feedparser==6.0.11

# Environment variables
python-dotenv==1.0.1

# HTTP requests
requests==2.32.3
